InfoDev Review (#228)
* InfoDev Review

Fixed brand names, except when in code or output messages. Technically the brand names should be fixed in output messages as well, but I wasn't sure if there are space constraints on the output messages so I didn't want to add words.
Corrected a few typos and fixed a broken link.
All changes to Intel brand names can be verified in namesdb.intel.com. All changes to 3rd party names (such as Torchvision) can be verified on their websites.

* Fixed typos flagged in review